MenuProject/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.7
sqlparse==0.4.3
